Crate slack_hook2

Library to send messages to slack rooms supports entire messaging API, including attachments and fields also support for built-in colors as well as any hex colors

Structs

Action
Actions are defined as an array, and values contained within it will be displayed with the message.

Attachment
Slack allows for attachments to be added to messages. See https://api.slack.com/docs/attachments for more information.

AttachmentBuilder
AttachmentBuilder is used to build a Attachment

Field
Fields are defined as an array, and hashes contained within it will be displayed in a table inside the message attachment.

HexColor
A HexColor String can be one of:

Payload
Payload to send to slack https://api.slack.com/incoming-webhooks https://api.slack.com/methods/chat.postMessage

PayloadBuilder
PayloadBuilder is used to build a Payload

Slack
Handles sending messages to slack

SlackLink
Representation of a link sent in slack

SlackText
Representation of any text sent through slack the text must be processed to escape specific characters

SlackTime
Slack timestamp

SlackUserLink
Representation of a user id link sent in slack

Enums

Parse
Change how messages are treated.

Section
Sections define parts of an attachment.

SlackColor
Default slack colors built-in to the API See: https://api.slack.com/docs/attachments

SlackError
Public error type

SlackTextContent
Enum used for constructing a text field having both SlackText(s) and SlackLink(s). The variants should be used together in a Vec on any function having a Into<SlackText> trait bound. The combined text will be space-separated.

Type Definitions

Result
Error handling convenience type